+++ bPlus/Node.java
@@ -2,7 +2,7 @@ package bPlus;
 
 public abstract class Node 
 {
-    public abstract long getFirstValue();
+    public abstract Integer getFirstValue();
 
     public abstract Node getPointerAt(int index);
